% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plus.R
\name{add_plus}
\alias{add_plus}
\title{Absolute Value Adding}
\usage{
add_plus(x, y)
}
\arguments{
\item{x}{Numeric. First adding number}

\item{y}{Numeric. Second adding number.}
}
\value{
Numeric. Sum of absolute values of inputs.
}
\description{
Absolute Value Adding
}
\examples{
add_plus(1,2)
add_plus(1,-2)
}
